requirements.txt
# requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
streamlit==1.28.0
streamlit-option-menu==0.3.6
pyyaml==6.0
joblib==1.3.2
scipy==1.11.4